requirements.txt
aioice==0.6.18
aiortc==0.9.27
av==7.0.1
cffi==1.14.0
crc32c==2.0
cryptography==2.8
JACK-Client==0.5.2
jedi==0.15.2
mypy==0.770
mypy-extensions==0.4.3
netifaces==0.10.9
parso==0.6.2
pluggy==0.13.1
pycparser==2.20
pyee==7.0.1
pylibsrtp==0.6.6
pyls==0.1.6
pyls-mypy==0.1.8
python-jsonrpc-server==0.3.4
python-language-server==0.31.9
six==1.14.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
ujson==1.35
websockets==8.1
